Update plan-upgrade-older-versions-office.md
Need to update the section 'Upgrade from Office 2016 or Office 2019 to Microsoft 365 Apps'. This section apply to any previous version of office (2016/2019/2021/2024/LTSC) not just 2016/2021. Add a comment on method1 using registry works only on C2R version not for office 2016 MSI. add on method2 that you need to remove previous installation if you need to upgrade from 32 bit to a 64 bits version.